backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
asgiref==3.11.0
    # via
    #   django
    #   django-cors-headers
    #   strawberry-graphql-django
beautifulsoup4==4.14.3
    # via yfinance
certifi==2026.1.4
    # via
    #   curl-cffi
    #   requests
cffi==2.0.0
    # via curl-cffi
charset-normalizer==3.4.4
    # via requests
cross-web==0.4.1
    # via strawberry-graphql
curl-cffi==0.13.0
    # via yfinance
django==5.2.10
    # via
    #   computational-finance (pyproject.toml)
    #   django-cors-headers
    #   strawberry-graphql-django
django-cors-headers==4.9.0
    # via computational-finance (pyproject.toml)
et-xmlfile==2.0.0
    # via openpyxl
frozendict==2.4.7
    # via yfinance
graphql-core==3.2.7
    # via strawberry-graphql
idna==3.11
    # via requests
multitasking==0.0.12
    # via yfinance
numpy==2.4.1
    # via
    #   computational-finance (pyproject.toml)
    #   pandas
    #   yfinance
openpyxl==3.1.5
    # via computational-finance (pyproject.toml)
packaging==26.0
    # via strawberry-graphql
pandas==3.0.0
    # via
    #   computational-finance (pyproject.toml)
    #   yfinance
peewee==3.19.0
    # via yfinance
platformdirs==4.5.1
    # via yfinance
protobuf==6.33.4
    # via yfinance
psycopg2-binary==2.9.11
    # via computational-finance (pyproject.toml)
pycparser==3.0
    # via cffi
python-dateutil==2.9.0.post0
    # via
    #   pandas
    #   strawberry-graphql
pytz==2025.2
    # via yfinance
requests==2.32.5
    # via
    #   computational-finance (pyproject.toml)
    #   yfinance
six==1.17.0
    # via python-dateutil
soupsieve==2.8.3
    # via beautifulsoup4
sqlparse==0.5.5
    # via django
strawberry-graphql==0.289.3
    # via strawberry-graphql-django
strawberry-graphql-django==0.74.1
    # via computational-finance (pyproject.toml)
tqdm==4.67.1
    # via computational-finance (pyproject.toml)
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   cross-web
    #   strawberry-graphql
urllib3==2.6.3
    # via requests
uuid6==2025.0.1
    # via computational-finance (pyproject.toml)
websockets==16.0
    # via yfinance
xlrd==2.0.2
    # via computational-finance (pyproject.toml)
yfinance==1.0
    # via computational-finance (pyproject.toml)
